+++ R/internal_functionality.R
@@ -259,3 +259,4 @@ setup_z <- function(y_vars, varAnnot) {
 
     z
 }
+