+++ tests/testthat/test-ui-radar.R
@@ -15,6 +15,8 @@ test_that("cXradar3", {
 
 test_that("cXradar4", {
     check_ui_test(cXradar4())
+
+    warning('Hover over any point - highlights middle circle point')
 })
 
 test_that("cXradar5", {